src/cloud_functions/job_source_scrapper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
google-cloud-bigquery==3.14.1
functions-framework==3.*